Deleting idea, pycache, and vscode directories. Renaming YSCFlask to Flask
