+++ Flask/static/Javascript/submitLogin.js
